Move instruction types
