+++ programs/token/src/definitions/approve.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
             MAX_MULTISIG_SIGNERS,
         },
@@ -8,7 +8,6 @@ use {
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
-    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -17,101 +16,86 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 6;
+const DISCRIMINATOR: u8 = 4;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - authority type (1 byte)
-///   - new authority (33 bytes, optional)
-const DATA_LEN: usize = 35;
+///   - amount (8 bytes)
+const DATA_LEN: usize = 9;
 
-#[repr(u8)]
-#[derive(Clone, Copy)]
-pub enum AuthorityType {
-    MintTokens = 0,
-    FreezeAccount = 1,
-    AccountOwner = 2,
-    CloseAccount = 3,
-}
-
-/// Sets a new authority of a mint or account.
+/// Approves a delegate.  A delegate is given the authority over tokens on
+/// behalf of the source account's owner.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single authority
-///   0. `[writable]` The mint or account to change the authority of.
-///   1. `[signer]` The current authority of the mint or account.
+///   * Single owner
+///   0. `[writable]` The source account.
+///   1. `[]` The delegate.
+///   2. `[signer]` The source account owner.
 ///
-///   * Multisignature authority
-///   0. `[writable]` The mint or account to change the authority of.
-///   1. `[]` The mint's or account's current multisignature authority.
-///   2. `..+M` `[signer]` M signer accounts.
-pub struct SetAuthority<
-    'account,
-    'address,
-    'multisig,
-    MultisigSigner: AsRef<AccountView>,
-    Program: TokenProgram,
-> {
-    /// The mint or account to change the authority of.
-    pub account: &'account AccountView,
+///   * Multisignature owner
+///   0. `[writable]` The source account.
+///   1. `[]` The delegate.
+///   2. `[]` The source account's multisignature owner.
+///   3. `..+M` `[signer]` M signer accounts.
+pub struct Approve<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
+    /// The source account.
+    pub source: &'account AccountView,
 
-    ///  The current authority of the mint or account.
+    /// The delegate.
+    pub delegate: &'account AccountView,
+
+    /// The source account owner.
     pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The type of authority to update.
-    pub authority_type: AuthorityType,
-
-    /// The new authority.
-    pub new_authority: Option<&'address Address>,
+    /// The amount of tokens the delegate is approved for.
+    pub amount: u64,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, 'address, Program: TokenProgram>
-    SetAuthority<'account, 'address, '_, &'account AccountView, Program>
-{
-    /// Creates a new `SetAuthority` instruction with a single authority.
+impl<'account, Program: TokenProgram> Approve<'account, '_, &'account AccountView, Program> {
+    /// Creates a new `Approve` instruction with a single owner authority.
     #[inline(always)]
     pub fn new(
-        account: &'account AccountView,
+        source: &'account AccountView,
+        delegate: &'account AccountView,
         authority: &'account AccountView,
-        authority_type: AuthorityType,
-        new_authority: Option<&'address Address>,
+        amount: u64,
     ) -> Self {
-        Self::with_multisig_signers(account, authority, authority_type, new_authority, &[])
+        Self::with_multisig_signers(source, delegate, authority, amount, &[])
     }
 }
 
-impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    SetAuthority<'account, 'address, 'multisig, MultisigSigner, Program>
+impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+    Approve<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `SetAuthority` instruction with a
-    /// multisignature authority and signer accounts.
+    /// Creates a new `Approve` instruction with a
+    /// multisignature owner authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        account: &'account AccountView,
+        source: &'account AccountView,
+        delegate: &'account AccountView,
         authority: &'account AccountView,
-        authority_type: AuthorityType,
-        new_authority: Option<&'address Address>,
+        amount: u64,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            account,
+            source,
+            delegate,
             authority,
             multisig_signers,
-            authority_type,
-            new_authority,
+            amount,
             _program: PhantomData,
         }
     }
@@ -140,7 +124,7 @@ impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program:
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -157,7 +141,7 @@ impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program:
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
+    for Approve<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -168,7 +152,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.account,
+            self.source,
+            self.delegate,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -184,7 +169,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.account,
+            self.source,
+            self.delegate,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -193,12 +179,12 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.authority_type, self.new_authority, data)
+        write_instruction_data(self.amount, data)
     }
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
+    for Approve<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
@@ -208,31 +194,29 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
     where
         Self: 'account + 'state,
     {
+        let Self {
+            source,
+            delegate,
+            authority,
+            multisig_signers,
+            amount,
+            ..
+        } = self;
+
         batch.push(
+            |accounts| write_accounts(source, delegate, authority, multisig_signers, accounts),
             |accounts| {
-                write_accounts(
-                    self.account,
-                    self.authority,
-                    self.multisig_signers,
-                    accounts,
-                )
+                write_instruction_accounts(source, delegate, authority, multisig_signers, accounts)
             },
-            |accounts| {
-                write_instruction_accounts(
-                    self.account,
-                    self.authority,
-                    self.multisig_signers,
-                    accounts,
-                )
-            },
-            |data| write_instruction_data(self.authority_type, self.new_authority, data),
+            |data| write_instruction_data(amount, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    account: &'account AccountView,
+    source: &'account AccountView,
+    delegate: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -241,21 +225,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if account.is_borrowed() {
+    if source.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(account, &mut accounts[0]);
+    CpiAccount::init_from_account_view(source, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
+    CpiAccount::init_from_account_view(delegate, &mut accounts[1]);
 
-    for (account, signer) in accounts[2..expected_accounts]
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
+
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -267,7 +253,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    account: &'account AccountView,
+    source: &'account AccountView,
+    delegate: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -276,21 +263,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(account.address()));
+    accounts[0].write(InstructionAccount::writable(source.address()));
 
-    accounts[1].write(InstructionAccount::new(
+    accounts[1].write(InstructionAccount::readonly(delegate.address()));
+
+    accounts[2].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[2..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -304,31 +293,16 @@ where
 
 #[inline(always)]
 fn write_instruction_data(
-    authority_type: AuthorityType,
-    new_authority: Option<&Address>,
+    amount: u64,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < 3 {
+    if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    data[1].write(authority_type as u8);
+    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
 
-    if let Some(new_authority) = new_authority {
-        if data.len() < DATA_LEN {
-            return Err(invalid_argument_error());
-        }
-
-        data[2].write(1);
-
-        write_bytes(&mut data[3..DATA_LEN], new_authority.as_array());
-
-        Ok(DATA_LEN)
-    } else {
-        data[2].write(0);
-
-        Ok(3)
-    }
+    Ok(DATA_LEN)
 }
