+++ programs/token/src/definitions/burn_checked.rs
@@ -1,10 +1,10 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
             MAX_MULTISIG_SIGNERS,
         },
-        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -16,67 +16,104 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 5;
+const DISCRIMINATOR: u8 = 15;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-const DATA_LEN: usize = 1;
+///   - amount (8 bytes)
+///   - decimals (1 byte)
+const DATA_LEN: usize = 10;
 
-/// Revokes the delegate's authority.
+/// Burns tokens by removing them from an account.
+/// [`super::BurnChecked`] does not support accounts
+/// associated with the native mint, use `CloseAccount` instead.
+///
+/// This instruction differs from Burn in that the decimals value is checked
+/// by the caller. This may be useful when creating transactions offline or
+/// within a hardware wallet.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner
-///   0. `[writable]` The source account.
-///   1. `[signer]` The source account's owner/delegate.
+///   * Single owner/delegate
+///   0. `[writable]` The account to burn from.
+///   1. `[writable]` The token mint.
+///   2. `[signer]` The account's owner/delegate.
 ///
 ///   * Multisignature owner/delegate
-///   0. `[writable]` The source account.
-///   1. `[]` The source account's multisignature owner/delegate.
-///   2. `..+M` `[signer]` M signer accounts.
-pub struct Revoke<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
-    /// The source account.
-    pub source: &'account AccountView,
+///   0. `[writable]` The account to burn from.
+///   1. `[writable]` The token mint.
+///   2. `[]` The account's multisignature owner/delegate.
+///   3. `..+M` `[signer]` M signer accounts.
+pub struct BurnChecked<
+    'account,
+    'multisig,
+    MultisigSigner: AsRef<AccountView>,
+    Program: TokenProgram,
+> {
+    /// The account to burn from.
+    pub account: &'account AccountView,
 
-    /// The source account's owner/delegate.
+    /// The token mint.
+    pub mint: &'account AccountView,
+
+    /// The account's owner/delegate.
     pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
+    /// The amount of tokens to burn.
+    pub amount: u64,
+
+    /// Expected number of base 10 digits to the right of the decimal place.
+    pub decimals: u8,
+
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> Revoke<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `Revoke` instruction with a single owner authority.
+impl<'account, Program: TokenProgram> BurnChecked<'account, '_, &'account AccountView, Program> {
+    /// Creates a new `BurnChecked` instruction with a single
+    /// owner/delegate authority.
     #[inline(always)]
-    pub fn new(source: &'account AccountView, authority: &'account AccountView) -> Self {
-        Self::with_multisig_signers(source, authority, &[])
+    pub fn new(
+        account: &'account AccountView,
+        mint: &'account AccountView,
+        authority: &'account AccountView,
+        amount: u64,
+        decimals: u8,
+    ) -> Self {
+        Self::with_multisig_signers(account, mint, authority, amount, decimals, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    Revoke<'account, 'multisig, MultisigSigner, Program>
+    BurnChecked<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `Revoke` instruction with a
-    /// multisignature owner authority and signer accounts.
+    /// Creates a new `BurnChecked` instruction with a
+    /// multisignature owner/delegate authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        source: &'account AccountView,
+        account: &'account AccountView,
+        mint: &'account AccountView,
         authority: &'account AccountView,
+        amount: u64,
+        decimals: u8,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            source,
+            account,
+            mint,
             authority,
             multisig_signers,
+            amount,
+            decimals,
             _program: PhantomData,
         }
     }
@@ -105,7 +142,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -122,7 +159,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for Revoke<'_, '_, MultisigSigner, Program>
+    for BurnChecked<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -132,7 +169,13 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_accounts(self.source, self.authority, self.multisig_signers, accounts)
+        write_accounts(
+            self.account,
+            self.mint,
+            self.authority,
+            self.multisig_signers,
+            accounts,
+        )
     }
 
     #[inline(always)]
@@ -143,17 +186,23 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(self.source, self.authority, self.multisig_signers, accounts)
+        write_instruction_accounts(
+            self.account,
+            self.mint,
+            self.authority,
+            self.multisig_signers,
+            accounts,
+        )
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(data)
+        write_instruction_data(self.amount, self.decimals, data)
     }
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for Revoke<'_, '_, MultisigSigner, Program>
+    for BurnChecked<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
@@ -164,23 +213,33 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| write_accounts(self.source, self.authority, self.multisig_signers, accounts),
             |accounts| {
-                write_instruction_accounts(
-                    self.source,
+                write_accounts(
+                    self.account,
+                    self.mint,
                     self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            write_instruction_data,
+            |accounts| {
+                write_instruction_accounts(
+                    self.account,
+                    self.mint,
+                    self.authority,
+                    self.multisig_signers,
+                    accounts,
+                )
+            },
+            |data| write_instruction_data(self.amount, self.decimals, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
+    account: &'account AccountView,
+    mint: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -189,21 +248,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if source.is_borrowed() {
+    if account.is_borrowed() | mint.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(source, &mut accounts[0]);
+    CpiAccount::init_from_account_view(account, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
+    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
 
-    for (account, signer) in accounts[2..expected_accounts]
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
+
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -215,7 +276,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
+    account: &'account AccountView,
+    mint: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -224,21 +286,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(source.address()));
+    accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::new(
+    accounts[1].write(InstructionAccount::writable(mint.address()));
+
+    accounts[2].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[2..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -251,12 +315,20 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
+fn write_instruction_data(
+    amount: u64,
+    decimals: u8,
+    data: &mut [MaybeUninit<u8>],
+) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
+
+    write_bytes(&mut data[1..9], &amount.to_le_bytes());
+
+    data[9].write(decimals);
 
     Ok(DATA_LEN)
 }
