+++ programs/token/src/definitions/initialize_account_2.rs
@@ -1,12 +1,13 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
         },
-        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_unchecked, CpiAccount},
         InstructionAccount, InstructionView,
@@ -15,62 +16,56 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 1;
+const DISCRIMINATOR: u8 = 16;
 
 /// Expected number of accounts.
-const ACCOUNTS_LEN: usize = 4;
+const ACCOUNTS_LEN: usize = 3;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-const DATA_LEN: usize = 1;
+///   - owner pubkey (32 bytes)
+const DATA_LEN: usize = 33;
 
-/// Initializes a new account to hold tokens.  If this account is associated
-/// with the native mint then the token balance of the initialized account
-/// will be equal to the amount of SOL in the account. If this account is
-/// associated with another mint, that mint must be initialized before this
-/// command can succeed.
-///
-/// The [`super::InitializeAccount`] instruction requires no
-/// signers and MUST be included within the same Transaction as the
-/// system program's `CreateAccount` instruction that creates the
-/// account being initialized. Otherwise another party can acquire
-/// ownership of the uninitialized account.
+/// Like [`super::InitializeAccount`], but the owner pubkey is
+/// passed via instruction data rather than the accounts list. This
+/// variant may be preferable when using Cross Program Invocation from
+/// an instruction that does not need the owner's `AccountInfo`
+/// otherwise.
 ///
 /// Accounts expected by this instruction:
 ///
 ///   0. `[writable]`  The account to initialize.
 ///   1. `[]` The mint this account will be associated with.
-///   2. `[]` The new account's owner/multisignature.
-///   3. `[]` Rent sysvar.
-pub struct InitializeAccount<'account, Program: TokenProgram> {
+///   2. `[]` Rent sysvar.
+pub struct InitializeAccount2<'account, Program: TokenProgram> {
     /// The account to initialize.
     pub account: &'account AccountView,
 
     /// The mint this account will be associated with.
     pub mint: &'account AccountView,
 
-    /// The new account's owner/multisignature.
-    pub owner: &'account AccountView,
-
     /// Rent sysvar.
     pub rent_sysvar: &'account AccountView,
+
+    /// The new account's owner/multisignature.
+    pub owner: &'account Address,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> InitializeAccount<'account, Program> {
+impl<'account, Program: TokenProgram> InitializeAccount2<'account, Program> {
     #[inline(always)]
     pub fn new(
         account: &'account AccountView,
         mint: &'account AccountView,
-        owner: &'account AccountView,
         rent_sysvar: &'account AccountView,
+        owner: &'account Address,
     ) -> Self {
         Self {
             account,
             mint,
-            owner,
             rent_sysvar,
+            owner,
             _program: PhantomData,
         }
     }
@@ -90,14 +85,14 @@ impl<'account, Program: TokenProgram> InitializeAccount<'account, Program> {
         unsafe {
             invoke_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
                     ),
                     data: from_raw_parts(instruction_data.as_ptr() as _, written_instruction_data),
                 },
-                from_raw_parts(accounts.as_ptr() as _, written_accounts),
+                from_raw_parts(accounts.as_ptr() as *const CpiAccount, written_accounts),
             );
         }
 
@@ -105,7 +100,7 @@ impl<'account, Program: TokenProgram> InitializeAccount<'account, Program> {
     }
 }
 
-impl<Program: TokenProgram> CpiWriter for InitializeAccount<'_, Program> {
+impl<Program: TokenProgram> CpiWriter for InitializeAccount2<'_, Program> {
     #[inline(always)]
     fn write_accounts<'cpi>(
         &self,
@@ -114,13 +109,7 @@ impl<Program: TokenProgram> CpiWriter for InitializeAccount<'_, Program> {
     where
         Self: 'cpi,
     {
-        write_accounts(
-            self.account,
-            self.mint,
-            self.owner,
-            self.rent_sysvar,
-            accounts,
-        )
+        write_accounts(self.account, self.mint, self.rent_sysvar, accounts)
     }
 
     #[inline(always)]
@@ -131,22 +120,16 @@ impl<Program: TokenProgram> CpiWriter for InitializeAccount<'_, Program> {
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(
-            self.account,
-            self.mint,
-            self.owner,
-            self.rent_sysvar,
-            accounts,
-        )
+        write_instruction_accounts(self.account, self.mint, self.rent_sysvar, accounts)
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(data)
+        write_instruction_data(self.owner, data)
     }
 }
 
-impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeAccount<'_, Program> {
+impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeAccount2<'_, Program> {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
@@ -156,25 +139,11 @@ impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeAccount<'_, 
         Self: 'account + 'state,
     {
         batch.push(
+            |accounts| write_accounts(self.account, self.mint, self.rent_sysvar, accounts),
             |accounts| {
-                write_accounts(
-                    self.account,
-                    self.mint,
-                    self.owner,
-                    self.rent_sysvar,
-                    accounts,
-                )
+                write_instruction_accounts(self.account, self.mint, self.rent_sysvar, accounts)
             },
-            |accounts| {
-                write_instruction_accounts(
-                    self.account,
-                    self.mint,
-                    self.owner,
-                    self.rent_sysvar,
-                    accounts,
-                )
-            },
-            write_instruction_data,
+            |data| write_instruction_data(self.owner, data),
         )
     }
 }
@@ -183,7 +152,6 @@ impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeAccount<'_, 
 fn write_accounts<'account, 'out>(
     account: &'account AccountView,
     mint: &'account AccountView,
-    owner: &'account AccountView,
     rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -202,9 +170,7 @@ where
 
     CpiAccount::init_from_account_view(mint, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(owner, &mut accounts[2]);
-
-    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[3]);
+    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[2]);
 
     Ok(ACCOUNTS_LEN)
 }
@@ -213,7 +179,6 @@ where
 fn write_instruction_accounts<'account, 'out>(
     account: &'account AccountView,
     mint: &'account AccountView,
-    owner: &'account AccountView,
     rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -228,20 +193,23 @@ where
 
     accounts[1].write(InstructionAccount::readonly(mint.address()));
 
-    accounts[2].write(InstructionAccount::readonly(owner.address()));
-
-    accounts[3].write(InstructionAccount::readonly(rent_sysvar.address()));
+    accounts[2].write(InstructionAccount::readonly(rent_sysvar.address()));
 
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
-fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
+fn write_instruction_data(
+    owner: &Address,
+    data: &mut [MaybeUninit<u8>],
+) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
+
+    write_bytes(&mut data[1..DATA_LEN], owner.as_array());
 
     Ok(DATA_LEN)
 }
