+++ programs/token/src/definitions/initialize_mint_2.rs
@@ -1,10 +1,13 @@
 use {
     crate::{
-        instructions::{invalid_argument_error, CpiWriter, TokenProgram},
+        definitions::{
+            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
+        },
         write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_unchecked, CpiAccount},
         InstructionAccount, InstructionView,
@@ -13,43 +16,53 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 23;
+const DISCRIMINATOR: u8 = 20;
 
 /// Expected number of accounts.
 const ACCOUNTS_LEN: usize = 1;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount (8 bytes)
-const DATA_LEN: usize = 9;
+///   - decimals (1 byte)
+///   - mint authority (32 bytes)
+///   - freeze authority (33 bytes, optional)
+const MAX_DATA_LEN: usize = 67;
 
-/// Convert an Amount of tokens to a `UiAmount` string, using the given
-/// mint.
-///
-/// Fails on an invalid mint.
-///
-/// Return data can be fetched using `sol_get_return_data` and deserialized
-/// with `String::from_utf8`.
+/// Like [`super::InitializeMint`], but does not require the Rent
+/// sysvar to be provided
 ///
 /// Accounts expected by this instruction:
 ///
-///   0. `[]` The mint to calculate for.
-pub struct AmountToUiAmount<'account, Program: TokenProgram> {
-    /// The mint to calculate for.
+///   0. `[writable]` The mint to initialize.
+pub struct InitializeMint2<'account, 'address, Program: TokenProgram> {
+    /// The mint to initialize.
     pub mint: &'account AccountView,
 
-    /// The amount of tokens to reformat.
-    pub amount: u64,
+    /// The number of base 10 digits to the right of the decimal place.
+    pub decimals: u8,
+
+    /// The authority/multisignature to mint tokens.
+    pub mint_authority: &'address Address,
+
+    /// The freeze authority/multisignature of the mint.
+    pub freeze_authority: Option<&'address Address>,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> AmountToUiAmount<'account, Program> {
+impl<'account, 'address, Program: TokenProgram> InitializeMint2<'account, 'address, Program> {
     #[inline(always)]
-    pub fn new(mint: &'account AccountView, amount: u64) -> Self {
+    pub fn new(
+        mint: &'account AccountView,
+        decimals: u8,
+        mint_authority: &'address Address,
+        freeze_authority: Option<&'address Address>,
+    ) -> Self {
         Self {
             mint,
-            amount,
+            decimals,
+            mint_authority,
+            freeze_authority,
             _program: PhantomData,
         }
     }
@@ -63,13 +76,13 @@ impl<'account, Program: TokenProgram> AmountToUiAmount<'account, Program> {
         let mut accounts = [UNINIT_CPI_ACCOUNT; ACCOUNTS_LEN];
         let written_accounts = self.write_accounts(&mut accounts)?;
 
-        let mut instruction_data = [UNINIT_BYTE; DATA_LEN];
+        let mut instruction_data = [UNINIT_BYTE; MAX_DATA_LEN];
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
             invoke_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -84,7 +97,7 @@ impl<'account, Program: TokenProgram> AmountToUiAmount<'account, Program> {
     }
 }
 
-impl<Program: TokenProgram> CpiWriter for AmountToUiAmount<'_, Program> {
+impl<Program: TokenProgram> CpiWriter for InitializeMint2<'_, '_, Program> {
     #[inline(always)]
     fn write_accounts<'cpi>(
         &self,
@@ -109,11 +122,16 @@ impl<Program: TokenProgram> CpiWriter for AmountToUiAmount<'_, Program> {
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, data)
+        write_instruction_data(
+            self.decimals,
+            self.mint_authority,
+            self.freeze_authority,
+            data,
+        )
     }
 }
 
-impl<Program: TokenProgram> super::IntoBatch<Program> for AmountToUiAmount<'_, Program> {
+impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeMint2<'_, '_, Program> {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
@@ -125,7 +143,14 @@ impl<Program: TokenProgram> super::IntoBatch<Program> for AmountToUiAmount<'_, P
         batch.push(
             |accounts| write_accounts(self.mint, accounts),
             |accounts| write_instruction_accounts(self.mint, accounts),
-            |data| write_instruction_data(self.amount, data),
+            |data| {
+                write_instruction_data(
+                    self.decimals,
+                    self.mint_authority,
+                    self.freeze_authority,
+                    data,
+                )
+            },
         )
     }
 }
@@ -140,6 +165,10 @@ where
 {
     if accounts.len() < ACCOUNTS_LEN {
         return Err(invalid_argument_error());
+    }
+
+    if mint.is_borrowed() {
+        return Err(account_borrow_failed_error());
     }
 
     CpiAccount::init_from_account_view(mint, &mut accounts[0]);
@@ -159,23 +188,37 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::readonly(mint.address()));
+    accounts[0].write(InstructionAccount::writable(mint.address()));
 
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
 fn write_instruction_data(
-    amount: u64,
+    decimals: u8,
+    mint_authority: &Address,
+    freeze_authority: Option<&Address>,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < DATA_LEN {
+    if data.len() < MAX_DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
+    data[1].write(decimals);
 
-    Ok(DATA_LEN)
+    write_bytes(&mut data[2..34], mint_authority.as_array());
+
+    if let Some(freeze_auth) = freeze_authority {
+        data[34].write(1);
+
+        write_bytes(&mut data[35..MAX_DATA_LEN], freeze_auth.as_array());
+
+        Ok(MAX_DATA_LEN)
+    } else {
+        data[34].write(0);
+
+        Ok(35)
+    }
 }
