+++ programs/token/src/definitions/initialize_multisig.rs
@@ -1,115 +1,104 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
     solana_instruction_view::{
-        cpi::{invoke_signed_unchecked, CpiAccount, Signer},
+        cpi::{invoke_unchecked, CpiAccount},
         InstructionAccount, InstructionView,
     },
     solana_program_error::{ProgramError, ProgramResult},
 };
 
+/// Maximum number of multisignature signers.
+pub const MAX_MULTISIG_SIGNERS: usize = 11;
+
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 8;
+const DISCRIMINATOR: u8 = 2;
 
 /// Maximum number of accounts expected by this instruction.
 ///
-/// The required number of accounts will depend whether the
-/// source account has a single owner or a multisignature
-/// owner.
-const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
+/// The required number of accounts will depend on the number of signer
+/// accounts.
+const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount (8 bytes)
-const DATA_LEN: usize = 9;
+///   - number of signers (1 byte)
+const DATA_LEN: usize = 2;
 
-/// Burns tokens by removing them from an account.  `Burn` does not support
-/// accounts associated with the native mint, use `CloseAccount` instead.
+/// Initializes a multisignature account with N provided signers.
+///
+/// Multisignature accounts can used in place of any single owner/delegate
+/// accounts in any token instruction that require an owner/delegate to be
+/// present.  The variant field represents the number of signers (M)
+/// required to validate this multisignature account.
+///
+/// The [`super::InitializeMultisig`] instruction requires no
+/// signers and MUST be included within the same Transaction as the
+/// system program's `CreateAccount` instruction that creates the
+/// account being initialized. Otherwise another party can acquire
+/// ownership of the uninitialized account.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner/delegate
-///   0. `[writable]` The account to burn from.
-///   1. `[writable]` The token mint.
-///   2. `[signer]` The account's owner/delegate.
-///
-///   * Multisignature owner/delegate
-///   0. `[writable]` The account to burn from.
-///   1. `[writable]` The token mint.
-///   2. `[]` The account's multisignature owner/delegate.
-///   3. `..+M` `[signer]` M signer accounts.
-pub struct Burn<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
-    /// The account to burn from.
-    pub account: &'account AccountView,
+///   0. `[writable]` The multisignature account to initialize.
+///   1. `[]` Rent sysvar.
+///   2. `..+N` `[signer]` The signer accounts, must equal to N where `1 <= N <=
+///      11`.
+pub struct InitializeMultisig<
+    'account,
+    'multisig,
+    MultisigSigner: AsRef<AccountView>,
+    Program: TokenProgram,
+> where
+    'account: 'multisig,
+{
+    /// The multisignature account to initialize.
+    pub multisig: &'account AccountView,
 
-    /// The token mint.
-    pub mint: &'account AccountView,
+    /// Rent sysvar.
+    pub rent_sysvar: &'account AccountView,
 
-    /// The account's owner/delegate.
-    pub authority: &'account AccountView,
-
-    /// Multisignature signers.
+    /// The signer accounts.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of tokens to burn.
-    pub amount: u64,
+    /// The number of signers (M) required to validate this multisignature
+    /// account.
+    pub m: u8,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> Burn<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `Burn` instruction with a single
-    /// owner/delegate authority.
+impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+    InitializeMultisig<'account, 'multisig, MultisigSigner, Program>
+where
+    'account: 'multisig,
+{
     #[inline(always)]
     pub fn new(
-        account: &'account AccountView,
-        mint: &'account AccountView,
-        authority: &'account AccountView,
-        amount: u64,
-    ) -> Self {
-        Self::with_multisig_signers(account, mint, authority, amount, &[])
-    }
-}
-
-impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    Burn<'account, 'multisig, MultisigSigner, Program>
-{
-    /// Creates a new `Burn` instruction with a
-    /// multisignature owner/delegate authority and signer accounts.
-    #[inline(always)]
-    pub fn with_multisig_signers(
-        account: &'account AccountView,
-        mint: &'account AccountView,
-        authority: &'account AccountView,
-        amount: u64,
+        multisig: &'account AccountView,
+        rent_sysvar: &'account AccountView,
         multisig_signers: &'multisig [MultisigSigner],
+        m: u8,
     ) -> Self {
         Self {
-            account,
-            mint,
-            authority,
+            multisig,
+            rent_sysvar,
             multisig_signers,
-            amount,
+            m,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
-    }
-
-    #[inline(always)]
-    pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
-            Err(ProgramError::InvalidArgument)?;
+            return Err(ProgramError::InvalidArgument);
         }
 
         let mut instruction_accounts = [UNINIT_INSTRUCTION_ACCOUNT; MAX_ACCOUNTS_LEN];
@@ -123,9 +112,9 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
-            invoke_signed_unchecked(
+            invoke_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -133,7 +122,6 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
                     data: from_raw_parts(instruction_data.as_ptr() as _, written_instruction_data),
                 },
                 from_raw_parts(accounts.as_ptr() as _, written_accounts),
-                signers,
             );
         }
 
@@ -142,7 +130,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for Burn<'_, '_, MultisigSigner, Program>
+    for InitializeMultisig<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -153,9 +141,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.account,
-            self.mint,
-            self.authority,
+            self.multisig,
+            self.rent_sysvar,
             self.multisig_signers,
             accounts,
         )
@@ -170,9 +157,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.account,
-            self.mint,
-            self.authority,
+            self.multisig,
+            self.rent_sysvar,
             self.multisig_signers,
             accounts,
         )
@@ -180,12 +166,12 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, data)
+        write_instruction_data(self.m, data)
     }
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for Burn<'_, '_, MultisigSigner, Program>
+    for InitializeMultisig<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
@@ -198,32 +184,29 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
-                    self.account,
-                    self.mint,
-                    self.authority,
+                    self.multisig,
+                    self.rent_sysvar,
                     self.multisig_signers,
                     accounts,
                 )
             },
             |accounts| {
                 write_instruction_accounts(
-                    self.account,
-                    self.mint,
-                    self.authority,
+                    self.multisig,
+                    self.rent_sysvar,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, data),
+            |data| write_instruction_data(self.m, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    account: &'account AccountView,
-    mint: &'account AccountView,
-    authority: &'account AccountView,
+    multisig: &'account AccountView,
+    rent_sysvar: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -231,23 +214,21 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if account.is_borrowed() | mint.is_borrowed() {
+    if multisig.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(account, &mut accounts[0]);
+    CpiAccount::init_from_account_view(multisig, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
+    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
-
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -259,9 +240,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    account: &'account AccountView,
-    mint: &'account AccountView,
-    authority: &'account AccountView,
+    multisig: &'account AccountView,
+    rent_sysvar: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -269,46 +249,35 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(account.address()));
+    accounts[0].write(InstructionAccount::writable(multisig.address()));
 
-    accounts[1].write(InstructionAccount::writable(mint.address()));
+    accounts[1].write(InstructionAccount::readonly(rent_sysvar.address()));
 
-    accounts[2].write(InstructionAccount::new(
-        authority.address(),
-        false,
-        multisig_signers.is_empty(),
-    ));
-
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
-        account.write(InstructionAccount::readonly_signer(
-            signer.as_ref().address(),
-        ));
+        account.write(InstructionAccount::readonly(signer.as_ref().address()));
     }
 
     Ok(expected_accounts)
 }
 
 #[inline(always)]
-fn write_instruction_data(
-    amount: u64,
-    data: &mut [MaybeUninit<u8>],
-) -> Result<usize, ProgramError> {
+fn write_instruction_data(m: u8, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
+    data[1].write(m);
 
     Ok(DATA_LEN)
 }
