+++ programs/token/src/definitions/revoke.rs
@@ -1,10 +1,10 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
             MAX_MULTISIG_SIGNERS,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -16,58 +16,34 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 12;
+const DISCRIMINATOR: u8 = 5;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 4 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount (8 bytes)
-///   - decimals (1 byte)
-const DATA_LEN: usize = 10;
+const DATA_LEN: usize = 1;
 
-/// Transfers tokens from one account to another either directly or via a
-/// delegate.  If this account is associated with the native mint then equal
-/// amounts of SOL and Tokens will be transferred to the destination
-/// account.
-///
-/// This instruction differs from [`super::Transfer`] in that the token mint and
-/// decimals value is checked by the caller.  This may be useful when
-/// creating transactions offline or within a hardware wallet.
+/// Revokes the delegate's authority.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner/delegate
+///   * Single owner
 ///   0. `[writable]` The source account.
-///   1. `[]` The token mint.
-///   2. `[writable]` The destination account.
-///   3. `[signer]` The source account's owner/delegate.
+///   1. `[signer]` The source account's owner/delegate.
 ///
 ///   * Multisignature owner/delegate
 ///   0. `[writable]` The source account.
-///   1. `[]` The token mint.
-///   2. `[writable]` The destination account.
-///   3. `[]` The source account's multisignature owner/delegate.
-///   4. `..+M` `[signer]` M signer accounts.
-pub struct TransferChecked<
-    'account,
-    'multisig,
-    MultisigSigner: AsRef<AccountView>,
-    Program: TokenProgram,
-> {
+///   1. `[]` The source account's multisignature owner/delegate.
+///   2. `..+M` `[signer]` M signer accounts.
+pub struct Revoke<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
     /// The source account.
-    pub from: &'account AccountView,
-
-    /// The token mint.
-    pub mint: &'account AccountView,
-
-    ///  The destination account.
-    pub to: &'account AccountView,
+    pub source: &'account AccountView,
 
     /// The source account's owner/delegate.
     pub authority: &'account AccountView,
@@ -75,56 +51,32 @@ pub struct TransferChecked<
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of tokens to transfer.
-    pub amount: u64,
-
-    /// Expected number of base 10 digits to the right of the decimal place.
-    pub decimals: u8,
-
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram>
-    TransferChecked<'account, '_, &'account AccountView, Program>
-{
-    /// Creates a new `TransferChecked` instruction with a single
-    /// owner/delegate authority.
+impl<'account, Program: TokenProgram> Revoke<'account, '_, &'account AccountView, Program> {
+    /// Creates a new `Revoke` instruction with a single owner authority.
     #[inline(always)]
-    pub fn new(
-        from: &'account AccountView,
-        mint: &'account AccountView,
-        to: &'account AccountView,
-        authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
-    ) -> Self {
-        Self::with_multisig_signers(from, mint, to, authority, amount, decimals, &[])
+    pub fn new(source: &'account AccountView, authority: &'account AccountView) -> Self {
+        Self::with_multisig_signers(source, authority, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    TransferChecked<'account, 'multisig, MultisigSigner, Program>
+    Revoke<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `TransferChecked` instruction with a
-    /// multisignature owner/delegate authority and signer accounts.
+    /// Creates a new `Revoke` instruction with a
+    /// multisignature owner authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        from: &'account AccountView,
-        mint: &'account AccountView,
-        to: &'account AccountView,
+        source: &'account AccountView,
         authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            from,
-            mint,
-            to,
+            source,
             authority,
             multisig_signers,
-            amount,
-            decimals,
             _program: PhantomData,
         }
     }
@@ -153,7 +105,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -170,7 +122,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for TransferChecked<'_, '_, MultisigSigner, Program>
+    for Revoke<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -180,14 +132,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_accounts(
-            self.from,
-            self.mint,
-            self.to,
-            self.authority,
-            self.multisig_signers,
-            accounts,
-        )
+        write_accounts(self.source, self.authority, self.multisig_signers, accounts)
     }
 
     #[inline(always)]
@@ -198,24 +143,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(
-            self.from,
-            self.mint,
-            self.to,
-            self.authority,
-            self.multisig_signers,
-            accounts,
-        )
+        write_instruction_accounts(self.source, self.authority, self.multisig_signers, accounts)
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, self.decimals, data)
+        write_instruction_data(data)
     }
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for TransferChecked<'_, '_, MultisigSigner, Program>
+    for Revoke<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
@@ -226,36 +164,23 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| {
-                write_accounts(
-                    self.from,
-                    self.mint,
-                    self.to,
-                    self.authority,
-                    self.multisig_signers,
-                    accounts,
-                )
-            },
+            |accounts| write_accounts(self.source, self.authority, self.multisig_signers, accounts),
             |accounts| {
                 write_instruction_accounts(
-                    self.from,
-                    self.mint,
-                    self.to,
+                    self.source,
                     self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, self.decimals, data),
+            write_instruction_data,
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    from: &'account AccountView,
-    mint: &'account AccountView,
-    to: &'account AccountView,
+    source: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -264,25 +189,21 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 4 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if from.is_borrowed() | to.is_borrowed() {
+    if source.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(from, &mut accounts[0]);
+    CpiAccount::init_from_account_view(source, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(to, &mut accounts[2]);
-
-    CpiAccount::init_from_account_view(authority, &mut accounts[3]);
-
-    for (account, signer) in accounts[4..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -294,9 +215,7 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    from: &'account AccountView,
-    mint: &'account AccountView,
-    to: &'account AccountView,
+    source: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -305,25 +224,21 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 4 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(from.address()));
+    accounts[0].write(InstructionAccount::writable(source.address()));
 
-    accounts[1].write(InstructionAccount::readonly(mint.address()));
-
-    accounts[2].write(InstructionAccount::writable(to.address()));
-
-    accounts[3].write(InstructionAccount::new(
+    accounts[1].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[4..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -336,20 +251,12 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(
-    amount: u64,
-    decimals: u8,
-    data: &mut [MaybeUninit<u8>],
-) -> Result<usize, ProgramError> {
+fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
-
-    write_bytes(&mut data[1..9], &amount.to_le_bytes());
-
-    data[9].write(decimals);
 
     Ok(DATA_LEN)
 }
