+++ programs/token/src/definitions/set_authority.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
             MAX_MULTISIG_SIGNERS,
         },
@@ -8,6 +8,7 @@ use {
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,86 +17,101 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 7;
+const DISCRIMINATOR: u8 = 6;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount to mint (8 bytes)
-const DATA_LEN: usize = 9;
+///   - authority type (1 byte)
+///   - new authority (33 bytes, optional)
+const DATA_LEN: usize = 35;
 
-/// Mints new tokens to an account.  The native mint does not support
-/// minting.
+#[repr(u8)]
+#[derive(Clone, Copy)]
+pub enum AuthorityType {
+    MintTokens = 0,
+    FreezeAccount = 1,
+    AccountOwner = 2,
+    CloseAccount = 3,
+}
+
+/// Sets a new authority of a mint or account.
 ///
 /// Accounts expected by this instruction:
 ///
 ///   * Single authority
-///   0. `[writable]` The mint.
-///   1. `[writable]` The account to mint tokens to.
-///   2. `[signer]` The mint's minting authority.
+///   0. `[writable]` The mint or account to change the authority of.
+///   1. `[signer]` The current authority of the mint or account.
 ///
 ///   * Multisignature authority
-///   0. `[writable]` The mint.
-///   1. `[writable]` The account to mint tokens to.
-///   2. `[]` The mint's multisignature mint-tokens authority.
-///   3. `..+M` `[signer]` M signer accounts.
-pub struct MintTo<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
-    /// The mint.
-    pub mint: &'account AccountView,
-
-    /// The account to mint tokens to.
+///   0. `[writable]` The mint or account to change the authority of.
+///   1. `[]` The mint's or account's current multisignature authority.
+///   2. `..+M` `[signer]` M signer accounts.
+pub struct SetAuthority<
+    'account,
+    'address,
+    'multisig,
+    MultisigSigner: AsRef<AccountView>,
+    Program: TokenProgram,
+> {
+    /// The mint or account to change the authority of.
     pub account: &'account AccountView,
 
-    /// The account to mint tokens to.
-    pub mint_authority: &'account AccountView,
+    ///  The current authority of the mint or account.
+    pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of new tokens to mint.
-    pub amount: u64,
+    /// The type of authority to update.
+    pub authority_type: AuthorityType,
+
+    /// The new authority.
+    pub new_authority: Option<&'address Address>,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> MintTo<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `MintTo` instruction with a single mint authority.
+impl<'account, 'address, Program: TokenProgram>
+    SetAuthority<'account, 'address, '_, &'account AccountView, Program>
+{
+    /// Creates a new `SetAuthority` instruction with a single authority.
     #[inline(always)]
     pub fn new(
-        mint: &'account AccountView,
         account: &'account AccountView,
-        mint_authority: &'account AccountView,
-        amount: u64,
+        authority: &'account AccountView,
+        authority_type: AuthorityType,
+        new_authority: Option<&'address Address>,
     ) -> Self {
-        Self::with_multisig_signers(mint, account, mint_authority, amount, &[])
+        Self::with_multisig_signers(account, authority, authority_type, new_authority, &[])
     }
 }
 
-impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    MintTo<'account, 'multisig, MultisigSigner, Program>
+impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+    SetAuthority<'account, 'address, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `MintTo` instruction with a
-    /// multisignature mint authority and signer accounts.
+    /// Creates a new `SetAuthority` instruction with a
+    /// multisignature authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        mint: &'account AccountView,
         account: &'account AccountView,
-        mint_authority: &'account AccountView,
-        amount: u64,
+        authority: &'account AccountView,
+        authority_type: AuthorityType,
+        new_authority: Option<&'address Address>,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            mint,
             account,
-            mint_authority,
+            authority,
             multisig_signers,
-            amount,
+            authority_type,
+            new_authority,
             _program: PhantomData,
         }
     }
@@ -124,7 +140,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -141,7 +157,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for MintTo<'_, '_, MultisigSigner, Program>
+    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -152,9 +168,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.mint,
             self.account,
-            self.mint_authority,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -169,9 +184,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.mint,
             self.account,
-            self.mint_authority,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -179,12 +193,12 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, data)
+        write_instruction_data(self.authority_type, self.new_authority, data)
     }
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for MintTo<'_, '_, MultisigSigner, Program>
+    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
@@ -197,32 +211,29 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
-                    self.mint,
                     self.account,
-                    self.mint_authority,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
             |accounts| {
                 write_instruction_accounts(
-                    self.mint,
                     self.account,
-                    self.mint_authority,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, data),
+            |data| write_instruction_data(self.authority_type, self.new_authority, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    mint: &'account AccountView,
     account: &'account AccountView,
-    mint_authority: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -230,23 +241,21 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if mint.is_borrowed() | account.is_borrowed() {
+    if account.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[0]);
+    CpiAccount::init_from_account_view(account, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(account, &mut accounts[1]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(mint_authority, &mut accounts[2]);
-
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -258,9 +267,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    mint: &'account AccountView,
     account: &'account AccountView,
-    mint_authority: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -268,23 +276,21 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(mint.address()));
+    accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::writable(account.address()));
-
-    accounts[2].write(InstructionAccount::new(
-        mint_authority.address(),
+    accounts[1].write(InstructionAccount::new(
+        authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -298,16 +304,31 @@ where
 
 #[inline(always)]
 fn write_instruction_data(
-    amount: u64,
+    authority_type: AuthorityType,
+    new_authority: Option<&Address>,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < DATA_LEN {
+    if data.len() < 3 {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
+    data[1].write(authority_type as u8);
 
-    Ok(DATA_LEN)
+    if let Some(new_authority) = new_authority {
+        if data.len() < DATA_LEN {
+            return Err(invalid_argument_error());
+        }
+
+        data[2].write(1);
+
+        write_bytes(&mut data[3..DATA_LEN], new_authority.as_array());
+
+        Ok(DATA_LEN)
+    } else {
+        data[2].write(0);
+
+        Ok(3)
+    }
 }
