+++ programs/token/src/definitions/unwrap_lamports.rs
@@ -1,10 +1,10 @@
 use {
     crate::{
-        instructions::{
+        definitions::{
             account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
             MAX_MULTISIG_SIGNERS,
         },
-        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -16,7 +16,7 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 10;
+const DISCRIMINATOR: u8 = 45;
 
 /// Maximum number of accounts expected by this instruction.
 ///
@@ -27,74 +27,97 @@ const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-const DATA_LEN: usize = 1;
+///   - amount (9 bytes, optional)
+const MAX_DATA_LEN: usize = 10;
 
-/// Freeze an Initialized account using the Mint's `freeze_authority` (if
-/// set).
+/// Enum specifying the amount of lamports to transfer
+/// from a native SOL account.
+#[repr(u8)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Amount {
+    /// Transfer the entire amount of the source account.
+    All,
+
+    /// Transfer a specified amount.
+    ///
+    /// The value must be less than or equal to the amount
+    /// of the source account.
+    Some(u64),
+}
+
+/// Transfer lamports from a native SOL account to a destination account.
+///
+/// This is useful to unwrap lamports from a wrapped SOL account.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner
-///   0. `[writable]` The account to freeze.
-///   1. `[]` The token mint.
-///   2. `[signer]` The mint freeze authority.
+///   * Single owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[writable]` The destination account.
+///   2. `[signer]` The source account's owner/delegate.
 ///
-///   * Multisignature owner
-///   0. `[writable]` The account to freeze.
-///   1. `[]` The token mint.
-///   2. `[]` The mint's multisignature freeze authority.
+///   * Multisignature owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[writable]` The destination account.
+///   2. `[]` The source account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
-pub struct FreezeAccount<
+pub struct UnwrapLamports<
     'account,
     'multisig,
     MultisigSigner: AsRef<AccountView>,
     Program: TokenProgram,
 > {
-    /// The account to freeze.
-    pub account: &'account AccountView,
+    /// The source account.
+    pub source: &'account AccountView,
 
-    /// The token mint.
-    pub mint: &'account AccountView,
+    /// The destination account.
+    pub destination: &'account AccountView,
 
-    /// The mint freeze authority.
-    pub freeze_authority: &'account AccountView,
+    /// The source account's owner/delegate.
+    pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
+    /// The amount of lamports to transfer.
+    pub amount: Amount,
+
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> FreezeAccount<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `FreezeAccount` instruction with a single freeze
+impl<'account, Program: TokenProgram> UnwrapLamports<'account, '_, &'account AccountView, Program> {
+    /// Creates a new `UnwrapLamports` instruction with a single owner
     /// authority.
     #[inline(always)]
     pub fn new(
-        account: &'account AccountView,
-        mint: &'account AccountView,
-        freeze_authority: &'account AccountView,
+        source: &'account AccountView,
+        destination: &'account AccountView,
+        authority: &'account AccountView,
+        amount: Amount,
     ) -> Self {
-        Self::with_multisig_signers(account, mint, freeze_authority, &[])
+        Self::with_multisig_signers(source, destination, authority, amount, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    FreezeAccount<'account, 'multisig, MultisigSigner, Program>
+    UnwrapLamports<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `FreezeAccount` instruction with a
-    /// multisignature freeze authority and signer accounts.
+    /// Creates a new `UnwrapLamports` instruction with a
+    /// multisignature owner authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        account: &'account AccountView,
-        mint: &'account AccountView,
-        freeze_authority: &'account AccountView,
+        source: &'account AccountView,
+        destination: &'account AccountView,
+        authority: &'account AccountView,
+        amount: Amount,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            account,
-            mint,
-            freeze_authority,
+            source,
+            destination,
+            authority,
             multisig_signers,
+            amount,
             _program: PhantomData,
         }
     }
@@ -117,13 +140,13 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         let mut accounts = [UNINIT_CPI_ACCOUNT; MAX_ACCOUNTS_LEN];
         let written_accounts = self.write_accounts(&mut accounts)?;
 
-        let mut instruction_data = [UNINIT_BYTE; DATA_LEN];
+        let mut instruction_data = [UNINIT_BYTE; MAX_DATA_LEN];
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::ID,
+                    program_id: &Program::id(),
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -140,7 +163,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for FreezeAccount<'_, '_, MultisigSigner, Program>
+    for UnwrapLamports<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -151,9 +174,9 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.account,
-            self.mint,
-            self.freeze_authority,
+            self.source,
+            self.destination,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -168,9 +191,9 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.account,
-            self.mint,
-            self.freeze_authority,
+            self.source,
+            self.destination,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -178,12 +201,12 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(data)
+        write_instruction_data(self.amount, data)
     }
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for FreezeAccount<'_, '_, MultisigSigner, Program>
+    for UnwrapLamports<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
@@ -196,32 +219,32 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
-                    self.account,
-                    self.mint,
-                    self.freeze_authority,
+                    self.source,
+                    self.destination,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
             |accounts| {
                 write_instruction_accounts(
-                    self.account,
-                    self.mint,
-                    self.freeze_authority,
+                    self.source,
+                    self.destination,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            write_instruction_data,
+            |data| write_instruction_data(self.amount, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    account: &'account AccountView,
-    mint: &'account AccountView,
-    freeze_authority: &'account AccountView,
+    source: &'account AccountView,
+    destination: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -235,15 +258,15 @@ where
         return Err(invalid_argument_error());
     }
 
-    if account.is_borrowed() {
+    if source.is_borrowed() | destination.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(account, &mut accounts[0]);
+    CpiAccount::init_from_account_view(source, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
+    CpiAccount::init_from_account_view(destination, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(freeze_authority, &mut accounts[2]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
 
     for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
@@ -257,9 +280,9 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    account: &'account AccountView,
-    mint: &'account AccountView,
-    freeze_authority: &'account AccountView,
+    source: &'account AccountView,
+    destination: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -273,12 +296,12 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(account.address()));
+    accounts[0].write(InstructionAccount::writable(source.address()));
 
-    accounts[1].write(InstructionAccount::readonly(mint.address()));
+    accounts[1].write(InstructionAccount::writable(destination.address()));
 
     accounts[2].write(InstructionAccount::new(
-        freeze_authority.address(),
+        authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
@@ -296,12 +319,25 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-    if data.len() < DATA_LEN {
+fn write_instruction_data(
+    amount: Amount,
+    data: &mut [MaybeUninit<u8>],
+) -> Result<usize, ProgramError> {
+    if data.len() < MAX_DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    Ok(DATA_LEN)
+    if let Amount::Some(amount) = amount {
+        data[1].write(1);
+
+        write_bytes(&mut data[2..MAX_DATA_LEN], &amount.to_le_bytes());
+
+        Ok(MAX_DATA_LEN)
+    } else {
+        data[1].write(0);
+
+        Ok(2)
+    }
 }
